requirements.txt
autopep8==1.6.0
mysql-connector-python==8.0.28
Pillow==9.1.0
protobuf==3.20.0
pycodestyle==2.8.0
toml==0.10.2
ttkbootstrap==1.7.4
